fix bugs when model_path endswith `/` or `\\`  (#565)
