+++ lazyllm/module/__init__.py
@@ -1,4 +1,5 @@
-from .module import ModuleBase, ActionModule, ServerModule, UrlModule, register
+from .module import ModuleBase, ActionModule, register
+from .servermodule import ServerModule, UrlModule
 from .trialmodule import TrialModule
 from .llms import (OnlineChatModule, OnlineChatModuleBase, OnlineEmbeddingModule,
                    OnlineEmbeddingModuleBase, AutoModel, TrainableModule)
